+++ libbb/u_signal_names.c
@@ -123,14 +123,14 @@ static const char signals[][7] = {
 
 int get_signum(const char *name)
 {
-	int i;
+	unsigned i;
 
 	i = bb_strtou(name, NULL, 10);
 	if (!errno)
 		return i;
 	if (strncasecmp(name, "SIG", 3) == 0)
 		name += 3;
-	for (i = 0; (size_t)i < ARRAY_SIZE(signals); i++)
+	for (i = 0; i < ARRAY_SIZE(signals); i++)
 		if (strcasecmp(name, signals[i]) == 0)
 			return i;
 
@@ -170,9 +170,9 @@ const char *get_signame(int number)
 
 void print_signames(void)
 {
-	int signo;
+	unsigned signo;
 
-	for (signo = 1; (size_t)signo < ARRAY_SIZE(signals); signo++) {
+	for (signo = 1; signo < ARRAY_SIZE(signals); signo++) {
 		const char *name = signals[signo];
 		if (name[0])
 			puts(name);
